requirements.txt
# Core Deep Learning Framework
torch>=1.12.0
torch-geometric>=2.0.0
torch-scatter>=2.0.9
torch-sparse>=0.6.15
torch-cluster>=1.6.0
torch-spline-conv>=1.2.1

# Graph Processing Libraries
networkx>=2.8.0
dgl>=1.0.0
scipy>=1.9.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Visualization
matplotlib>=3.5.0
plotly>=5.10.0
seaborn>=0.11.0

# Jupyter and Colab
jupyter>=1.0.0
ipywidgets>=7.7.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
kaggle>=1.5.12

# Optional: For advanced features
optuna>=3.0.0  # Hyperparameter optimization
wandb>=0.13.0  # Experiment tracking
